pyproject.toml
[tool.poetry]
name = "django-task#2"
version = "0.1.0"
description = ""
authors = ["AdhamAliAbdelAal <adhamali20012015@gmail.com>"]
readme = "README.md"
packages = [{include = "django_task#2"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
